requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
androguard==4.1.2
androidviewclient==23.0.1
annotated-types==0.6.0
anyio==4.3.0
apkInspector==1.2.1
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
banal==1.0.6
certifi==2024.2.2
Chameleon==4.5.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
culebratester-client==2.0.64
cycler==0.12.1
dataset==1.6.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.48.1
frida==16.1.11
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
ipython==8.21.0
javaobj-py3==0.4.4
jedi==0.19.1
kiwisolver==1.4.5
loguru==0.7.2
lxml==5.1.0
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
multidict==6.0.5
mutf8==1.0.6
networkx==3.2.1
numpy==1.26.4
openai==1.13.3
oscrypto==1.3.0
packaging==23.2
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pyaxml==0.0.5
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.12.0
sniffio==1.3.0
SQLAlchemy==1.4.51
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.7.1
urllib3==2.2.0
wcwidth==0.2.13
win32-setctime==1.1.0
yarl==1.9.4

setuptools~=65.5.0
screeninfo~=0.8.1